restructure into game subdir
